+++ common/util.py
@@ -295,6 +295,7 @@ ACCESS_FREE = "Free"
 ACCESS_PRO = "Pro"
 ACCESS_PREMIUM = "Premium"
 ACCESS_TESTER = "Tester"
+ACCESS_TRIAL = "Trial"
 ACCESS_SUPER = "Super"
 
 # STATES