+++ jtool.py
@@ -224,6 +224,13 @@ class JTool():
         "Terminus Machina",
         "Chaos Infinity"
       ]
+    elif self.get_level() == util.ACCESS_TRIAL:
+      LIST_DUNGEON = [
+        "Hazardous Valley (Awakened)",
+        "Steamer Crazy (Awakened)",
+        "Holy Keldrasil",
+        "Chaos Infinity"
+      ]
     elif self.get_level() == util.ACCESS_SUPER:
       LIST_DUNGEON = [
         "Hazardous Valley (Awakened)",
@@ -248,12 +255,12 @@ class JTool():
       case util.DATA_DUNGEON:
         return util.STATE_NORMAL
       case util.DATA_CONNECTION:
-        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO or self.get_level() == util.ACCESS_TESTER:
+        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO or self.get_level() == util.ACCESS_TESTER or self.get_level() == util.ACCESS_TRIAL:
           return util.STATE_DISABLED
         else:
           return util.STATE_NORMAL
       case util.DATA_PET:
-        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO or self.get_level() == util.ACCESS_PREMIUM:
+        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO or self.get_level() == util.ACCESS_PREMIUM or self.get_level() == util.ACCESS_TRIAL:
           return util.STATE_DISABLED
         else:
           return util.STATE_NORMAL
@@ -282,7 +289,7 @@ class JTool():
       case util.DATA_RANGE:
         return util.STATE_NORMAL
       case util.DATA_VERADRIX:
-        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO:
+        if self.get_level() == util.ACCESS_FREE or self.get_level() == util.ACCESS_PRO or self.get_level() == util.ACCESS_TRIAL:
           return util.STATE_DISABLED
         else:
           return util.STATE_NORMAL
@@ -565,7 +572,7 @@ class JTool():
     tab_pricing = ttk.Frame(tab_control)
     tab_control.add(tab_dungeon, text=util.TAB_DUNGEON, state=self.get_access(util.DATA_DUNGEON))
     tab_control.add(tab_connection, text=util.TAB_CONNECTION, state=self.get_access(util.DATA_CONNECTION))
-    tab_control.add(tab_pet, text=util.TAB_PET, state=self.get_access(util.DATA_CONNECTION))
+    tab_control.add(tab_pet, text=util.TAB_PET, state=self.get_access(util.DATA_PET))
     tab_control.add(tab_misc, text=util.TAB_OTHERS, state=self.get_access(util.DATA_OTHERS))
     tab_control.add(tab_pricing, text=util.TAB_PRICING)
     tab_control.pack(expand=1, fill="both")